requirements.txt
Gooey==1.0.8.1
google-api-python-client==2.127.0
google-auth-oauthlib==1.2.0
python-dotenv~=1.0.1